feat/fe-main : import 순서, 변수명 등 정리
