+++ client/src/pages/RegisterPage/styles.ts
@@ -74,41 +74,3 @@ export const idValidationStyle = (isAllValid: string) =>
     color: isAllValid === RESULT.SUCCESS ? 'green' : 'red',
     fontSize: FONT_SIZE.SMALL,
   });
-
-export const paginationStyle = css({
-  fontSize: FONT_SIZE.MEDIUM,
-  color: COLORS.TEXT_1,
-  '& .pagination': {
-    display: 'flex',
-    justifyContent: 'center',
-    marginTop: 10,
-  },
-  // '& ul': {
-  //   listStyle: 'none',
-  //   padding: 0,
-  // },
-  '& ul.pagination li': {
-    width: 30,
-    height: 30,
-    border: '1px solid #e2e2e2',
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  '& ul.pagination li:first-child, ul.pagination li:last-child': {
-    borderRadius: '5px 0 0 5px',
-  },
-  '& ul.pagination li a': {
-    textDecoration: 'none',
-    color: COLORS.PRIMARY_1,
-  },
-  '& ul.pagination li.active a': {
-    color: 'white',
-  },
-  '& ul.pagination li.active': {
-    backgroundColor: COLORS.PRIMARY_2,
-  },
-  '& ul.pagination li a:hover, ul.pagination li a.active': {
-    color: COLORS.TEXT_1,
-  },
-});
